Added Saplo API wrapper
All Saplo API calls now go though a javascript wrapper layer to improve
readability of the code. This also improves error handling and decreases number
of Saplo API calls we need to make.

As an effect of this we now return Url.id as response to 'tagga' so a similar
change needs to be deployed in scrape.js and the webbackend.